Implemented AppContextProvider
