+++ src/components/column/styles/Column.css
@@ -5,10 +5,6 @@
     border-radius: 2px;
     margin-right: 4.5px;
     background-color: #fff;
-
-    /* border: 1px solid #dadde2; */
-    /* border-radius: 10px; */
-    /* border-radius: 4px; */
     border-radius: 6px;
 }
 
@@ -20,12 +16,9 @@
     padding-bottom: 12px;
     text-align: center;
     border-bottom: 2px solid #f2f4f7;
-    /* border-bottom: 2px solid #e3e5e9; */
+    border-bottom: 2px solid #f2f3f5;
 }
 
-.column-header {
-    /* width: 80%; */
-}
 .column-header:hover {
     cursor: pointer;
 }
