+++ src/components/column/styles/ColumnAddTemplate.css
@@ -7,7 +7,7 @@
     /* border: 1px dashed #a8aaaf; */
     border: 1px dashed #b0b2b7;
     margin-top: 1%;
-    margin-left: 0.1%;
+    margin-left: 0.2%;
 }
 .add-column-template--container:hover {
     cursor: pointer;
