+++ src/components/features/error/ErrorMessage.css
@@ -3,16 +3,20 @@
   margin: 0 auto;
   width: 88%;
   height: 18px;
+  height: 16px;
   border-radius: 10px;
   padding: 20px;
   display: flex;
   margin-bottom: 6px;
   margin-top: -2px;
+  padding-top: 16px;
 }
 
 .err-msg--container {
-    border: 1px solid #e83232;
-    background-color: #e83232;
+    /* border: 1px solid #e83232; */
+    border: 1px solid #ef3434;
+    /* background-color: #e83232; */
+    background-color: #ef3434;
 }
 
 .err-msg {
