+++ src/components/task/Task.css
@@ -9,9 +9,11 @@
 }
 .task--container:hover { 
     cursor: pointer !important;
-    background-color: #f7f8f9 !important;
-    background-color: #fafafc !important;
+    /* background-color: #fafafc !important;
+    background-color: #f7f8fa !important; */
+    background-color: #f9f9fa !important;
     border: 1px solid #d2d9e1 !important;
+    /* border: 1px solid #d0d7df !important; */
 }
 
 .task-title {
