+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import { AppProvider } from './AppProvider';
+import { AppContextProvider } from './AppContextProvider';
 import reportWebVitals from './reportWebVitals';
 import { BrowserRouter } from 'react-router-dom'
 import './index.css';
@@ -11,9 +11,9 @@ const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <BrowserRouter>
-			<AppProvider>
+			<AppContextProvider>
 				<App />
-			</AppProvider>
+			</AppContextProvider>
 		</BrowserRouter>
   </React.StrictMode>
 );
